# cmd/internal/features/tasks/transport/http/get_task.go
package tasks_transport_http

import (
	"net/http"

	core_logger "github.com/catstyle1101/todo_app_go/cmd/internal/core/logger"
	core_http_request "github.com/catstyle1101/todo_app_go/cmd/internal/core/transport/http/request"
	core_http_response "github.com/catstyle1101/todo_app_go/cmd/internal/core/transport/http/response"
)

type GetTaskResponse TaskDtoResponse

// GetTask 		godoc
// @Summary 	Получить задачу
// @Description Получение конкретной задачи по ее ID
// @Tags 		tasks
// @Produce 	json
// @Param 		task_id 	path 	int 	true 				"ID получаемой задачи"
// @Success 	200 		{object} 		GetTaskResponse 					"Задача успешно найдена"
// @Failure 	400 		{object} 		core_http_response.ErrorResponse 	"Bad request"
// @Failure 	404 		{object} 		core_http_response.ErrorResponse 	"Task not found"
// @Failure 	500 		{object} 		core_http_response.ErrorResponse 	"Internal Server Error"
// @Router 		/tasks/{task_id} 			[get]
func (h *TaskHTTPHandler) GetTask(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	taskID, err := core_http_request.GetPathIntValue(r, "task_id")
	if err != nil {
		responseHandler.ErrorResponse(err, "get task_id path value")
		return
	}

	taskDomain, err := h.taskService.GetTask(ctx, taskID)
	if err != nil {
		responseHandler.ErrorResponse(err, "failed to get task")
		return
	}

	response := GetTaskResponse(taskDTOFromDomain(taskDomain))

	responseHandler.JSONResponse(response, http.StatusOK)

}